refactor: 타입넣는중
