+++ api/share.ts
@@ -1,16 +1,36 @@
+type User = {
+  id: number;
+  name: string;
+  email: string;
+  profileImageSource: string;
+};
+
+type getUserResponse = {
+  data: User;
+};
+
 const baseUrl = new URL("https://bootcamp-api.codeit.kr");
 const getUrl = (path: string) => new URL(path, baseUrl);
 
 export const getShareUserData = async () => {
   const requestUrl = getUrl("/api/sample/user");
-  const response = await fetch(requestUrl);
-  const jsonData = await response.json();
-  return [response, jsonData];
+  const response = await fetch(requestUrl, {
+    method: "GET",
+  });
+
+  if (response.status === 200) {
+    const jsonData = await response.json();
+    return jsonData as Promise<getUserResponse>;
+  }
 };
 
-export const getShareData = async () => {
+export const getShareFolderData = async () => {
   const requestUrl = getUrl("/api/sample/folder");
-  const response = await fetch(requestUrl);
-  const jsonData = await response.json();
-  return [response, jsonData];
+  const response = await fetch(requestUrl, {
+    method: "GET",
+  });
+  if (response.status === 200) {
+    const jsonData = await response.json();
+    return jsonData;
+  }
 };
