+++ pages/shared/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import useFetchData from "@/hooks/useFetchData";
-import { getShareData } from "@/api/share";
+import { getShareFolderData } from "@/api/share";
 import ShareNav from "@/components/nav/ShareNav";
 import Header from "@/components/header/Header";
 import SearchBar from "@/components/searchbar/SearchBar";
@@ -8,7 +8,7 @@ import ImageList from "@/components/imagelist/ImageList";
 import Footer from "@/components/footer/Footer";
 
 export default function SharedPage() {
-  const [data, isLoading] = useFetchData(getShareData);
+  const [data, isLoading] = useFetchData(getShareFolderData);
 
   return (
     !isLoading && (
